+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Please follow [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## Unreleased
+
+### Added
+
+- Add `Fracindex::from_hex` for decoding hexadecimal index representations.
+
 ## v0.1.0 (2026-07-30)
 
 ### Added